Refactor code to share more logic between thunderbird and browsers
